+++ packages/block-library/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New Features
 
--   Navigation: Add Space around and Space evenly justification options ([#71129](https://github.com/WordPress/gutenberg/pull/71129)).
+-   Buttons and Navigation: Add Space around and Space evenly justification options ([#71129](https://github.com/WordPress/gutenberg/pull/71129)).
 
 ### Enhancements